+++ Modules/Intent.Modules.FluentValidation.Shared/NuGetPackages.cs
@@ -5,5 +5,4 @@ namespace Intent.Modules.FluentValidation.Shared;
 public static class NuGetPackages
 {
     public static readonly INugetPackageInfo FluentValidation = new NugetPackageInfo("FluentValidation", "11.6.0");
-    public static readonly INugetPackageInfo FluentValidationDependencyInjectionExtensions = new NugetPackageInfo("FluentValidation.DependencyInjectionExtensions", "11.6.0");
 }